requirements.txt
Flask==3.1.0
firebase-admin==6.6.0
python-dotenv==1.1.0
